requirements.txt
transformers>=4.30.0
datasets>=2.10.0
torch>=1.13.0
bitsandbytes>=0.39.0
scikit-learn
PyYAML
peft
accelerate